Refactor and harden audio and model utilities
- Normalize formatting across app and library modules
- Add safer audio stream handling and cleanup config imports
- Tighten a few runtime checks and exception paths

+++ lib/uvr5_pack/demucs/filtering.py
@@ -1,8 +1,6 @@
 from typing import Optional
 import torch
-import torch.nn as nn
-from torch import Tensor
-from torch.utils.data import DataLoader
+
 
 def atan2(y, x):
     r"""Element-wise arctangent function of y/x.
@@ -45,7 +43,9 @@ def _norm(x: torch.Tensor) -> torch.Tensor:
     return torch.abs(x[..., 0]) ** 2 + torch.abs(x[..., 1]) ** 2
 
 
-def _mul_add(a: torch.Tensor, b: torch.Tensor, out: Optional[torch.Tensor] = None) -> torch.Tensor:
+def _mul_add(
+    a: torch.Tensor, b: torch.Tensor, out: Optional[torch.Tensor] = None
+) -> torch.Tensor:
     """Element-wise multiplication of two complex Tensors described
     through their real and imaginary parts.
     The result is added to the `out` tensor"""
@@ -64,7 +64,9 @@ def _mul_add(a: torch.Tensor, b: torch.Tensor, out: Optional[torch.Tensor] = Non
     return out
 
 
-def _mul(a: torch.Tensor, b: torch.Tensor, out: Optional[torch.Tensor] = None) -> torch.Tensor:
+def _mul(
+    a: torch.Tensor, b: torch.Tensor, out: Optional[torch.Tensor] = None
+) -> torch.Tensor:
     """Element-wise multiplication of two complex Tensors described
     through their real and imaginary parts
     can work in place in case out is a only"""
@@ -259,18 +261,24 @@ def expectation_maximization(
 
     # allocate the spatial covariance matrices
     R = [
-        torch.zeros((nb_bins, nb_channels, nb_channels, 2), dtype=x.dtype, device=x.device)
+        torch.zeros(
+            (nb_bins, nb_channels, nb_channels, 2), dtype=x.dtype, device=x.device
+        )
         for j in range(nb_sources)
     ]
     weight: torch.Tensor = torch.zeros((nb_bins,), dtype=x.dtype, device=x.device)
 
-    v: torch.Tensor = torch.zeros((nb_frames, nb_bins, nb_sources), dtype=x.dtype, device=x.device)
+    v: torch.Tensor = torch.zeros(
+        (nb_frames, nb_bins, nb_sources), dtype=x.dtype, device=x.device
+    )
     for it in range(iterations):
         # constructing the mixture covariance matrix. Doing it with a loop
         # to avoid storing anytime in RAM the whole 6D tensor
 
         # update the PSD as the average spectrogram over channels
-        v = torch.mean(torch.abs(y[..., 0, :]) ** 2 + torch.abs(y[..., 1, :]) ** 2, dim=-2)
+        v = torch.mean(
+            torch.abs(y[..., 0, :]) ** 2 + torch.abs(y[..., 1, :]) ** 2, dim=-2
+        )
 
         # update spatial covariance matrices (weighted update)
         for j in range(nb_sources):
@@ -308,7 +316,6 @@ def expectation_maximization(
 
             # separate the sources
             for j in range(nb_sources):
-
                 # create a wiener gain for this source
                 gain = torch.zeros_like(inv_Cxx)
 
@@ -328,7 +335,9 @@ def expectation_maximization(
 
                 # apply it to the mixture
                 for i in range(nb_channels):
-                    y[t, ..., j] = _mul_add(gain[..., i, :], x[t, ..., i, None, :], y[t, ..., j])
+                    y[t, ..., j] = _mul_add(
+                        gain[..., i, :], x[t, ..., i, None, :], y[t, ..., j]
+                    )
 
     return y, v, R
 
@@ -434,7 +443,12 @@ def wiener(
             mix_stft[..., None]
             * (
                 targets_spectrograms
-                / (eps + torch.sum(targets_spectrograms, dim=-1, keepdim=True).to(mix_stft.dtype))
+                / (
+                    eps
+                    + torch.sum(targets_spectrograms, dim=-1, keepdim=True).to(
+                        mix_stft.dtype
+                    )
+                )
             )[..., None, :]
         )
     else:
